Prevent multiple calls to -flush.
Flushes are performed as background tasks, but since the task ID
is stored as an instance variable, calling -flush twice in a row
obliterates the first task ID and allows only the second one to
be ended correctly. This results in iOS killing your application
because a background task has been running for too long.