Expose notifications, update via API
Similar to how Survey's are exposed through the iOS SDK, this change
also adds the ability to retrieve all available notifications.

It also adds the ability to manually check for updated notifications,
surveys, etc. Depending on when the Mixpanel instance is initialized
(e.g. well after the app is activated) this may be necessary.