Initial Commit of NSSecureCoding Implementation
This implementation, I believe, conforms to the "letter of the law" of NSSecureCoding, and makes progress on the "spirit" of the law.

I'm implementing this to reduce warnings about lack of NSSecureCoding and to investigate a possible fix to a NSCoding crash I'm seeing on iOS 15.1 beta.

Someone who understands the implementation better should probably clamp down more on MPTweakValue given that it seems that may just be intended to store a basic type.